requirements.txt
streamlit
pandas
scikit-learn
nltk
Pillow
Pytesseract